# reader_concurrency_semaphore: Optimize resource_units destruction by postponing wait list processing