@paszkow 2024-12-13
Replaces https://github.com/scylladb/scylladb/pull/21346