@scylladb-promoter 2024-12-13
#### :red_circle: CI State: FAILURE

:white_check_mark: - [Build](https://jenkins.scylladb.com/job/scylla-master/job/scylla-ci/13812/)
:x: - [Unit Tests Custom](https://jenkins.scylladb.com/job/scylla-master/job/scylla-ci/13812/)
**The following new/updated tests ran 100 times for each mode:**
:small_blue_diamond: boost/reader_concurrency_semaphore_test
#### Failed Tests (2/27):
- [test_reader_concurrency_semaphore_admission](https://jenkins.scylladb.com//job/scylla-master/job/scylla-ci/13812/testReport/junit/boost/reader_concurrency_semaphore_test/test_reader_concurrency_semaphore_admission) [:mag:](https://github.com/scylladb/scylladb/issues?q=is:issue+is:open+test_reader_concurrency_semaphore_admission)
- [test_reader_concurrency_semaphore_necessary_evicting](https://jenkins.scylladb.com//job/scylla-master/job/scylla-ci/13812/testReport/junit/boost/reader_concurrency_semaphore_test/test_reader_concurrency_semaphore_necessary_evicting) [:mag:](https://github.com/scylladb/scylladb/issues?q=is:issue+is:open+test_reader_concurrency_semaphore_necessary_evicting)
#### Build Details:
- Duration: 27 min
- Builder: spider2.cloudius-systems.com