@paszkow 2024-12-20
> ❌ - [Offline-installer Artifact Tests](https://jenkins.scylladb.com/job/scylla-master/job/scylla-ci-offline-installer/175/)

Reported errors:

```
Error occured during collecting of db.crt on host: artifacts-debian11-jenkins-db-node-557c9825-0-1
'NoneType' object is not iterable
Error occured during collecting of client-facing.crt on host: artifacts-debian11-jenkins-db-node-557c9825-0-1
'NoneType' object is not iterable
Error occured during collecting of cqlshrc on host: artifacts-debian11-jenkins-db-node-557c9825-0-1
'NoneType' object is not iterable
```

> #### Failed Tests (1/36625):
> * [test_mv_topology_change.dev.3](https://jenkins.scylladb.com//job/scylla-master/job/scylla-ci/13927/testReport/junit/topology_custom/test_mv_topology_change/Tests___Unit_Tests___test_mv_topology_change_dev_3) [🔍](https://github.com/scylladb/scylladb/issues?q=is:issue+is:open+test_mv_topology_change.dev.3)
> 
Existing issue: https://github.com/scylladb/scylladb/issues/21912