@mykaul 2024-12-22
Unit tests failed with https://github.com/scylladb/scylladb/issues/21912

Offline installer with some debian 11 issue which I have no idea what it is - and its history has vanished.