@avikivity 2024-12-22
I asked for before/after perf-simple-query report (customary is with --smp 1 -m 1G) so we see effects on the fast-path too.