@paszkow 2024-12-23
> I asked for before/after perf-simple-query report (customary is with --smp 1 -m 1G) so we see effects on the fast-path too.

@avikivity  `perf-simple-query --tablets --smp 1 -m 1G` results:

Before:
```
enable-cache=1
Running test with config: {partitions=10000, concurrency=100, mode=read, frontend=cql, query_single_key=no, counters=no}
Disabling auto compaction
Creating 10000 partitions...
251338.33 tps ( 63.1 allocs/op,   0.0 logallocs/op,  14.1 tasks/op,   41343 insns/op,   18467 cycles/op,        0 errors)
271951.42 tps ( 63.1 allocs/op,   0.0 logallocs/op,  14.1 tasks/op,   41343 insns/op,   17507 cycles/op,        0 errors)
271525.13 tps ( 63.1 allocs/op,   0.0 logallocs/op,  14.1 tasks/op,   41330 insns/op,   17522 cycles/op,        0 errors)
272440.88 tps ( 63.1 allocs/op,   0.0 logallocs/op,  14.1 tasks/op,   41339 insns/op,   17477 cycles/op,        0 errors)
277071.85 tps ( 63.0 allocs/op,   0.0 logallocs/op,  14.1 tasks/op,   41341 insns/op,   17179 cycles/op,        0 errors)

throughput: mean=268865.52 standard-deviation=10048.95 median=271951.42 median-absolute-deviation=3575.35 maximum=277071.85 minimum=251338.33
instructions_per_op: mean=41339.31 standard-deviation=5.47 median=41341.14 median-absolute-deviation=3.35 maximum=41343.44 minimum=41329.91
cpu_cycles_per_op: mean=17630.26 standard-deviation=488.62 median=17506.64 median-absolute-deviation=153.16 maximum=18467.24 minimum=17178.67
```

After:
```
enable-cache=1
Running test with config: {partitions=10000, concurrency=100, mode=read, frontend=cql, query_single_key=no, counters=no}
Disabling auto compaction
Creating 10000 partitions...

255880.51 tps ( 63.1 allocs/op,   0.0 logallocs/op,  14.1 tasks/op,   41030 insns/op,   18351 cycles/op,        0 errors)
253799.46 tps ( 63.1 allocs/op,   0.0 logallocs/op,  14.1 tasks/op,   41024 insns/op,   18724 cycles/op,        0 errors)
249800.55 tps ( 63.1 allocs/op,   0.0 logallocs/op,  14.1 tasks/op,   41024 insns/op,   18971 cycles/op,        0 errors)
263477.75 tps ( 63.1 allocs/op,   0.0 logallocs/op,  14.1 tasks/op,   41013 insns/op,   18015 cycles/op,        0 errors)
264105.07 tps ( 63.1 allocs/op,   0.0 logallocs/op,  14.1 tasks/op,   41023 insns/op,   18017 cycles/op,        0 errors)

throughput: mean=257412.66 standard-deviation=6223.35 median=255880.51 median-absolute-deviation=6065.08 maximum=264105.07 minimum=249800.55
instructions_per_op: mean=41022.98 standard-deviation=6.10 median=41023.74 median-absolute-deviation=1.23 maximum=41030.21 minimum=41013.28
cpu_cycles_per_op: mean=18415.82 standard-deviation=426.32 median=18351.46 median-absolute-deviation=398.80 maximum=18971.20 minimum=18015.36

```